Castling (#2)
* Comment cleanup

Old method that is no longer used and is not needed for reference

* Castling execution implemented

* IsAttackedBy() started

This is needed to look for checks, checkmates, and castling availability

* 5/22 Cleanup

Square.IsAttackedBy() started, not finished yet
Warning message fixes

* ChessboardBackend.Initialize() added

This supports future feature to create a deep copy of backend in order to check move validity

* Enable deep copying Move

* Enable deep copying Coordinate

* MoveDoesNotPutOwnKingInCheck implemented

Some debug tools were added in this commit as well. It is now possible to see if a move would cause your own king to be in check.

* refactor: Maintainability pass

Moved MoveDoesNotPutOwnKingInCheck() to Move.cs from ChessboardBackend.cs
various Console.WriteLine ()statements edited
various whitespace changes for readability

* Move that puts king in check cannot be executed

+ a bunch of comment changes for UX

* Queenside castle fixed

queenside castle moves king 2 spaces, not 3 spaces

* improved messages

* simplified logic

* removed deprecated method and method calls